% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chandwich.R
\docType{data}
\name{owtemps}
\alias{owtemps}
\title{Oxford and Worthing annual maximum temperatures}
\format{A dataframe with 80 rows and 2 columns, named Oxford and Worthing.}
\source{
Tabony, R. C. (1983) Extreme value analysis in meteorology.
 \emph{The Meteorological Magazine}, \strong{112}, 77-98.
}
\usage{
owtemps
}
\description{
Annual maximum temperatures at Oxford and Worthing (England), for the
period 1901 to 1980.
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{https://doi.org/10.1093/biomet/asm015}
}
\keyword{datasets}
