% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.confint}
\alias{plot.confint}
\title{Plot diagnostics a confint object}
\usage{
\method{plot}{confint}(x, y, ..., add_lines = TRUE)
}
\arguments{
\item{x}{an object of class "confint", a result of a call to
\code{\link{conf_intervals}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to ...}
}
\description{
\code{plot} method for class "confint".
}
