% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.confreg}
\alias{plot.confreg}
\title{Plot diagnostics a confreg object}
\usage{
\method{plot}{confreg}(x, y = NULL, y2 = NULL, y3 = NULL, conf = 95,
  ...)
}
\arguments{
\item{x}{an object of class "confreg", a result of a call to
\code{\link{conf_region}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed to \code{\link[graphics]{contour}}.}
}
\description{
\code{plot} method for class "confreg".
}
