% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_print.R
\name{print.compmod}
\alias{print.compmod}
\title{Print method for objects of class "compmod"}
\usage{
\method{print}{compmod}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{an object of class "compmod", a result of a call to
\code{\link{compare_models}}.}

\item{digits}{An integer. The argument \code{digits} to
\code{\link[base:Round]{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
  methods.
}
\description{
\code{print} method for class "compmod".
}
\details{
Prints the name of the model, the null (H0) and alternative
hypotheses (HA), the test statistic, degrees of freedom and the p-value.
If the test is based on the approximation detailed by equations (18)-(20)
of Chandler and Bate (2007), rather than equation (17), then this stated.
}
\section{Examples}{

See the examples in \code{\link{compare_models}}.
}

\seealso{
\code{\link{adjust_loglik}} to adjust a user-supplied
  loglikelihood function.

\code{\link{compare_models}} to compare nested models using an
  (adjusted) likelihood ratio test.
}
