% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{print.confint}
\alias{print.confint}
\title{Print method for objects of class "confint"}
\usage{
\method{print}{confint}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "confint", a result of a call to
\code{\link{conf_intervals}}.}

\item{digits}{An integer. The argument \code{digits} to
\code{\link{print.default}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
The argument \code{x}, invisibly, as for all \code{\link{print}}
  methods.
}
\description{
\code{print} method for class "confint".
}
\details{
Prints the name of the model, details of any fixed parameters,
  the confidence level of the interval(s) and whether or not
  the loglikelihood has been adjusted, and symmetric and (profile) likelihood
  based intervals.
}
\section{Examples}{

See the examples in \code{\link{conf_intervals}}.
}

\seealso{
\code{\link{adjust_loglik}} to adjust a user-supplied
  loglikelhood function.

\code{\link{conf_intervals}} for confidence intervals for
  individual parameters.
}
