% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary_print.R
\name{print.summary.chandwich}
\alias{print.summary.chandwich}
\title{Print method for objects of class "summary.chandwich"}
\usage{
\method{print}{summary.chandwich}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.chandwich", a result of a call to
\code{\link{summary.chandwich}}.}

\item{...}{Additional arguments passed on to \code{print}.}
}
\value{
Prints a numeric matrix with 3 columns and the number of rows
  equal to the number of parameters in the current model, i.e.
  \code{attr(object, "p_current")}.
  The columns contain: the maximum likelihood estimates (MLE), unadjusted
  standard errors (SE) and adjusted standard errors (adjSE).
}
\description{
\code{print} method for an object \code{object} of class "summary.chandwich".
}
\section{Examples}{

See the examples in \code{\link{adjust_loglik}}.
}

\seealso{
\code{\link{adjust_loglik}} to adjust a user-supplied
  loglikelihood.

\code{\link{summary.chandwich}} for a diagnostic plot.
}
