% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{profile_loglik}
\alias{profile_loglik}
\title{Profile loglikelihood}
\usage{
profile_loglik(object, prof_pars = NULL, prof_vals = NULL, init = NULL,
  type = c("vertical", "cholesky", "dilation", "none"), ...)
}
\arguments{
\item{object}{An object of class \code{"chandwich"} returned by
\code{adjust_loglik}.}

\item{prof_pars}{A vector specifying the subset of the (unfixed) parameters
  over which to profile. Can be either a numeric vector, specifying indices
  of the components of the \strong{full} parameter vector, or a character
  vector of parameter names, which must be a subset of those supplied in
  \code{par_names} in the call to \code{\link{adjust_loglik}} that produced
  \code{object}.

  \code{prof_pars} must not have any parameters in common with
  \code{attr(object, "fixed_pars")}.  \code{prof_pars} must not contain
  all of the unfixed parameters, i.e. there is no point in profiling over
  all the unfixed parameters.}

\item{prof_vals}{A numeric vector.  Values of the parameters in
\code{prof_pars}.  If \code{prof_vals = NULL} then the MLEs stored
in \code{object} of the parameters \code{prof_pars} are used.}

\item{init}{A numeric vector of initial estimates of the values of the
parameters that are not fixed and are not in \code{prof_pars}.
Should have length \code{attr(object, "p_current") - length(prof_pars)}.
If \code{init} is \code{NULL} or is of the wrong length then the
relevant components from the MLE stored in \code{object} are used.}

\item{type}{A character scalar.  The argument \code{type} to the function
returned by \code{\link{adjust_loglik}}, that is, the type of adjustment
made to the independence loglikelihood function.}

\item{...}{Further arguments to be passed to \code{\link[stats]{optim}}.
These may include \code{gr}, \code{method}, \code{lower}, \code{upper}
or \code{control}.}
}
\value{
A numeric vector of length 1.  The value of the profile
  loglikelihood.
}
\description{
Calculates the profile loglikelihood for a subset of the model parameters.
}
\examples{
profile_loglik(large, prof_pars = "xi1")
}
