% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chandwich.R
\docType{data}
\name{rats}
\alias{rats}
\title{Rat tumor data}
\format{A matrix with 71 rows and 2 columns.}
\source{
Table 5.1 of Gelman, A., Carlin, J. B., Stern, H. S. Dunson, D. B.,
 Vehtari, A. and Rubin, D. B. (2013) \emph{Bayesian Data Analysis},
 Chapman & Hall / CRC.
  \url{http://www.stat.columbia.edu/~gelman/book/data/rats.asc}
}
\usage{
rats
}
\description{
Tumor incidence in 71 groups of rate from Tarone (1982).
The matrix \code{rat} has 71 rows and 2 columns.
Each row relates to a different group of rats.
The first column (\code{y}) contains the number of rats with tumors.
The second column (\code{n}) contains the total number of rats.
}
\references{
Tarone, R. E. (1982) The use of historical information in
  testing for a trend in proportions. \emph{Biometrics}, \strong{38},
  215-220. \url{https://www.jstor.org/stable/2530304}
}
\keyword{datasets}
