% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{summary.chandwich}
\alias{summary.chandwich}
\title{Summarizing adjusted loglikelihoods}
\usage{
\method{summary}{chandwich}(object, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{object}{an object of class "chandwich", a result of a call to
\code{adjust_loglik}.}

\item{digits}{A integer. Used for number formatting with
\code{\link{signif}}.}

\item{...}{Additional optional arguments. At present no optional
arguments are used.}
}
\value{
Returns a numeric matrix with 3 or 4 columns and \code{object$d}
  rows, where \code{object$d} is the number of parameters in the model
  defined by the argument \code{loglik} in the call to
  \code{\link{adjust_loglik}}.
  The columns contain: the parameter names (if these were supplied using
  \code{par_names} in the call to \code{\link{adjust_loglik}}),
  the MLE, unadjusted standard errors (SE) and adjusted standard errors
  (adjSE).
}
\description{
\code{summary} method for class "chandwich"
}
\examples{
norm_loglik <- function(params, data) {
  mu <- params[1]
  sigma <- params[2]
  if (sigma <= 0) {
    return(-Inf)
  }
  return(dnorm(data, mean = mu, sd = sigma, log = TRUE))
}
mu <- 0
sigma <- 1
norm_data <- rnorm(2000, mean = mu, sd = sigma)
mu_sigma <- c(0, 1)
cluster <- 1:length(norm_data)
cluster <- rep(1:40, 50)

pjn <- adjust_loglik(loglik = norm_loglik, data = norm_data, cluster = cluster,
              init = 0:1)
summary(pjn)
}
\seealso{
\code{\link{adjust_loglik}} to adjust a user-supplied
  loglikelhood function.

\code{\link{plot.chandwich}} for plots of one-dimensional adjusted
  loglikelihoods.

\code{\link{adjust_loglik}}.
}
