% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_loglik.R
\name{adjust_loglik}
\alias{adjust_loglik}
\title{Loglikelihood adjustment}
\usage{
adjust_loglik(loglik, ..., cluster, d = 1, init = 0.1)
}
\arguments{
\item{loglik}{An R function.  The first argument must be the vector of
model parameter(s).  Returns a vector - use the length of this to check
that cluster is the correct length!}

\item{...}{Further arguments to be passed to logf.}

\item{cluster}{A numeric vector.  Must have length \code{length(data)} (if
\code{data} is a vector) or \code{nrow(data)} (if \code{data} is a matrix.}
}
\value{
A function that evaluates the adjusted loglikelihood.
   + ARGUMENTS horizontal1, horizontal2, vertical
}
\description{
Chandler-Bate
}
\examples{
binom_loglik <- function(prob, data) {
  return(dbinom(data[, "x"], data[, "size"], prob, log = TRUE))
}
prob <- 0.1
binom_data <- cbind(size = 10, x = rbinom(20, 10, prob))
binom_loglik(prob, binom_data)
cluster <- rep(1, nrow(binom_data))
adjust_loglik(loglik = binom_loglik, data = binom_data, cluster = cluster)
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
