% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chandwich.R
\docType{package}
\name{chandwich}
\alias{chandwich}
\alias{chandwich-package}
\title{chandwich: Chandler-Bate Sandwich Estimator Loglikelihood Adjustment}
\description{
Performs adjustments of the independence loglikelihood using
a robust sandwich estimator of the parameter covariance matrix, based on
the methodology in
\href{http://dx.doi.org/10.1093/biomet/asm015}{Chandler and Beta (2007)}.
This can be used for cluster correlated data when interest lies in the
parameters of the marginal distributions.
}
\details{
Add details

  See \code{vignette("chandwich-vignette", package = "chandwich")} for an
  overview of the package.
}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
