% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Comparison of nested models}
\usage{
compare_models(larger, smaller = NULL, approx = FALSE, fixed_pars = NULL,
  fixed_at = 0, init = NULL, ...)
}
\arguments{
\item{larger}{An object of class \code{"chandwich"} returned by
\code{adjust_loglik}.  The larger of the two models.}

\item{smaller}{An object of class \code{"chandwich"} returned by
\code{adjust_loglik}.  The smaller of the two models.}

\item{approx}{A logical scalar.  If \code{approx = TRUE} then the
approximation detailed by equations (18)-(20) of Chandler and Bates(2007)
is used.  This option is only used if \code{smaller} is supplied.
If \code{approx = FALSE} then the adjusted likelihood is
maximised under the restriction imposed by \code{delta}.}

\item{fixed_pars}{A numeric vector.}

\item{fixed_at}{A numeric vector.}

\item{init}{A numeric vector of initial values for use in the search for
the MLE under the smaller model.}

\item{...}{Further arguments to be passed to \code{\link[stats]{optim}}.
These may include \code{gr}, \code{method}, \code{lower}, \code{upper}
or \code{control}.}
}
\value{
Test statistic, degrees of freedom (q) and p-value.
}
\description{
adjusted likelihood ratio statistics (ALRS),
following Section 3.5 of
\href{http://dx.doi.org/10.1093/biomet/asm015}{Chandler and Bate (2007)}.
}
\details{
add details
}
\examples{
# GEV model, owtemps data ----------
# ... following Section 5.2 of Chandler and Bate (2007)

gev_loglik <- function(pars, data) {
  o_pars <- pars[c(1, 3, 5)] + pars[c(2, 4, 6)]
  w_pars <- pars[c(1, 3, 5)] - pars[c(2, 4, 6)]
  if (o_pars[2] <= 0 | w_pars[2] <= 0) return(-Inf)
  o_loglik <- gev_dens(data[, "Oxford"], o_pars[1], o_pars[2], o_pars[3],
                       log = TRUE)
  w_loglik <- gev_dens(data[, "Worthing"], w_pars[1], w_pars[2], w_pars[3],
                       log = TRUE)
  return(o_loglik + w_loglik)
}

# Initial estimates (method of moments for the Gumbel case)
sigma <- as.numeric(sqrt(6 * diag(stats::var(owtemps))) / pi)
mu <- as.numeric(colMeans(owtemps) - 0.57722 * sigma)
init <- c(mean(mu), -diff(mu) / 2, mean(sigma), -diff(sigma) / 2, 0, 0)

# Perform the log-likelihood adjustment
large <- adjust_loglik(gev_loglik, data = owtemps, init = init,
         par_names = c("mu0", "mu1", "sigma0", "sigma1", "xi0", "xi1"))
# Rows 1, 3 and 4 of Table 2 of Chandler and Bate (2007)
round(attr(large, "MLE"), 4)
round(attr(large, "SE"), 4)
round(attr(large, "adjSE"), 4)

# Fix xi1 = 0
medium <- adjust_loglik(larger = large, fixed_pars = 6)
# Fix sigma1 = xi1 = 0
small <- adjust_loglik(larger = medium, fixed_pars = c(4, 6))

# Test xi1 = 0 (2 equivalent ways)
compare_models(large, fixed_pars = 6)$p_value
compare_models(large, medium)$p_value

# Test xi1 = 0, using approximation
compare_models(large, medium, approx = TRUE)$p_value

# Test sigma1 = 0 for xi1 = 0
compare_models(medium, small)$p_value

# Test sigma1 = xi1 = 0
compare_models(large, small)$p_value

}
\references{
Chandler, R. E. and Bate, S. (2007). Inference for clustered
  data using the independence loglikelihood. \emph{Biometrika},
  \strong{94}(1), 167-183. \url{http://dx.doi.org/10.1093/biomet/asm015}
}
\seealso{
\code{\link{adjust_loglik}}: to adjust a user-supplied
  loglikelhood function.

\code{\link{summary.chandwich}} for maximum likelihood estimates
  and unadjusted and adjusted standard errors.

\code{\link{plot.chandwich}} for one- and two- dimensional plots
  of of adjusted loglikelihoods.
}
