% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_summary.R
\name{plot.chandwich}
\alias{plot.chandwich}
\title{Plot diagnostics a chandwich object}
\usage{
\method{plot}{chandwich}(x, y, ..., lower = NULL, upper = NULL, type = 1)
}
\arguments{
\item{x}{an object of class "chandwich", a result of a call to
\code{\link{adjust_loglik}}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed on to ...}

\item{lower, upper}{Numeric vectors specifying the lower and upper limits
on the parameter values to appear in the plot.  If either \code{lower}
or \code{upper} are not provided then the MLE minus (for \code{lower})
or plus (for \code{upper}) three (adjusted) standard errors is used.}

\item{type}{An integer vector, a subset of the numbers \code{1:4}.
Indicates which loglikelihoods to plot: \code{1} for \code{"vertical"}
adjustment; \code{2} for \code{"cholesky"} (horizontal adjustment);
\code{3} for \code{"dilation"} (horizontal adjustment); \code{4}
for no adjustment, i.e. based on the independence loglikelihood.}
}
\value{
A list containing the graphical parameters using in producing the
  plot including any arguments supplied via ... is returned (invisibly).
}
\description{
\code{plot} method for class "chandwich".
}
\examples{
binom_loglik <- function(prob, data) {
  if (prob < 0 || prob > 1) {
    return(-Inf)
  }
  return(dbinom(data[, "y"], data[, "n"], prob, log = TRUE))
}
cluster <- 1:nrow(rats)
rat_res <- adjust_loglik(loglik = binom_loglik, data = rats, cluster = cluster)
plot(rat_res)
plot(rat_res, type = 1:4)
plot(rat_res, type = 1:4, lower = 0, upper = 1)
}
\seealso{
\code{\link{adjust_loglik}}.

\code{\link{summary.chandwich}} for maximum likelihood estimates
  and unadjusted and adjusted standard errors.
}
